# backport-2.1: distsqlrun: bugfixes/perf change to joinReader 