Backport 1/1 commits from #28980.

/cc @cockroachdb/release

---

Previously, joinReader used pretty-printed string keys to correlate its
input rows with its looked-up rows. This was incorrect and inefficient.

Now, it uses bytes keys like everything else. Also, fix a bug that
caused partial key creation from EncDatums on tables with interleaves to
panic and/or return incorrect results.

Closes #28981.

```
name                      old time/op    new time/op     delta
JoinReader/rows=16-24        134µs ± 1%       57µs ± 1%   -57.39%  (p=0.000 n=9+10)
JoinReader/rows=256-24      1.54ms ± 1%     0.42ms ± 1%   -72.43%  (p=0.000 n=10+9)
JoinReader/rows=4096-24     45.4ms ± 1%     24.5ms ± 1%   -46.14%  (p=0.000 n=8+10)
JoinReader/rows=65536-24     5.03s ± 1%      4.80s ± 1%    -4.64%  (p=0.000 n=9+10)

name                      old speed      new speed       delta
JoinReader/rows=16-24     2.87MB/s ± 1%   6.74MB/s ± 1%  +134.76%  (p=0.000 n=7+10)
JoinReader/rows=256-24    4.00MB/s ± 1%  14.52MB/s ± 1%  +262.83%  (p=0.000 n=10+10)
JoinReader/rows=4096-24   2.17MB/s ± 1%   4.02MB/s ± 1%   +85.64%  (p=0.000 n=8+10)
JoinReader/rows=65536-24   310kB/s ± 0%    327kB/s ± 2%    +5.48%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op    delta
JoinReader/rows=16-24       40.4kB ± 1%     18.1kB ± 1%   -55.08%  (p=0.000 n=10+10)
JoinReader/rows=256-24       390kB ± 0%       93kB ± 1%   -76.07%  (p=0.000 n=8+9)
JoinReader/rows=4096-24     5.96MB ± 0%     1.25MB ± 1%   -78.99%  (p=0.000 n=9+9)
JoinReader/rows=65536-24    99.2MB ± 6%     23.2MB ±15%   -76.66%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op   delta
JoinReader/rows=16-24          646 ± 0%        179 ± 0%   -72.37%  (p=0.000 n=10+10)
JoinReader/rows=256-24       8.77k ± 0%      1.33k ± 0%   -84.83%  (p=0.000 n=10+9)
JoinReader/rows=4096-24       139k ± 0%        20k ± 0%   -85.48%  (p=0.000 n=9+9)
JoinReader/rows=65536-24     2.24M ± 1%      0.33M ± 7%   -85.14%  (p=0.000 n=9+9)
```